Added super janky debugging
Hey guys this is an absolutely terrible way to debug, but it was easy
to implement and it works using the Tester.  Just cat file.txt to see
what happened and delete after each run.  The first things I’m noticing
is that our seqno’s never increment and the very first packet is always
resent